The Timeline of Work Update
Finally committing the timeline section of the portfolio. Took so long and there is still so much to do...

This will include new logos and images for the timeline. A new CSS page, changes to the other CSS pages, javascript updates, changes to the main page. Just a lot of everything.