Added logo to navbar
Added my logo next to my name in the navbar. This required me to change the default font size for the starting name but still looks good. There are minor issues with different screen sizes that will need to be fixed later. Full screen and mobile are good for now.